#79 feat : daily, week 공통 컴포넌트 추출
